pyproject.toml
[tool.poetry]
name = "trackmate-instructions"
version = "0.1.0"
description = ""
authors = ["Oliver Hofmann"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.42.0"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
